- Add a Graph using LiveChart that can show the function of Cartesian System - Change the type of data of functions to Decimal to be able to store more percise and bigger range of numeral values - Add Tangant Function to be read.

+++ CPP/FormulaParse.cs
@@ -212,17 +212,17 @@ namespace CPP
                     // Values
                     case "p":
                     case "P":
-                        bt.InsertSingleNode(root, new SingleNode(root, 3.14));
+                        bt.InsertSingleNode(root, new SingleNode(root, 3.14m));
                         break;
 
                     //Real and Natural Numbers
                     case "r":
                     case "n":
-                        root = bt.InsertSingleNode(root, new SingleNode(root, Convert.ToDouble(input[++i])));
+                        root = bt.InsertSingleNode(root, new SingleNode(root, Convert.ToDecimal(input[++i])));
                         break;
 
                     default:
-                        bt.InsertSingleNode(root, new SingleNode(root, Convert.ToDouble(input[i])));
+                        bt.InsertSingleNode(root, new SingleNode(root, Convert.ToDecimal(input[i])));
                         break;
                 }
